+++ javascript/components/Camera.js
@@ -6,8 +6,6 @@ import locationManager from '../modules/location/locationManager';
 import {isNumber, toJSONString, viewPropTypes, existenceChange} from '../utils';
 import * as geoUtils from '../utils/geoUtils';
 
-import NativeBridgeComponent from './NativeBridgeComponent';
-
 const MapboxGL = NativeModules.MGLModule;
 
 export const NATIVE_MODULE_NAME = 'RCTMGLCamera';
@@ -67,7 +65,7 @@ const SettingsPropTypes = {
   zoomLevel: PropTypes.number,
 };
 
-class Camera extends NativeBridgeComponent {
+class Camera extends React.Component {
   static propTypes = {
     ...viewPropTypes,